Change the criteria used for splitting a message in blocks with RSA
In the subject, when a message is to be sent and is specified as
a sequence of letters, the criterion to split seems to be whether
the size of the message in characters is larger than the block
size for the given n, regardless of whether the numeric
representation of the message is smaller than n or not.

However, when the signature is computed from the rubric, the
criterion seems to simply be whether the numeric representation
of the rubric is smaller than n.

Since the purpose of this library is to be useful for students of
the subject, this commit matches that behaviour.